re-organize common rpc data types
