+++ src/rpc_client/common_ops.rs
@@ -1,11 +1,11 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use crate::lotus_json::lotus_json_with_self;
-use crate::rpc_api::{common_api::*, data_types::APIVersion};
+use crate::rpc_api::{
+    common_api::{DISCOVER, SESSION, SHUTDOWN, START_TIME, VERSION},
+    data_types::{APIVersion, DiscoverResult},
+};
 use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
 
 use super::{ApiInfo, JsonRpcError, RpcRequest};
 
@@ -38,37 +38,3 @@ impl ApiInfo {
         RpcRequest::new(SESSION, ())
     }
 }
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct DiscoverResult {
-    info: DiscoverInfo,
-    methods: Vec<DiscoverMethod>,
-    openrpc: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DiscoverMethod {
-    deprecated: bool,
-    description: String,
-    external_docs: DiscoverDocs,
-    name: String,
-    param_structure: String,
-    params: Value,
-    // result
-    summary: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct DiscoverDocs {
-    description: String,
-    url: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct DiscoverInfo {
-    title: String,
-    version: String,
-}
-
-lotus_json_with_self!(DiscoverResult, DiscoverMethod, DiscoverDocs, DiscoverInfo);
